fix: update strictNullChecks to true in tsconfig.json
This commit updates the tsconfig.json file to set the strictNullChecks option to true, ensuring stricter null checking in the TypeScript codebase.

It also makes the description, capital, and passwordConfirmation properties optional in the TownDto, DepartmentDto, and CreateUserDto classes respectively. This change allows for more flexibility when working with these DTOs.

These changes were made to improve code quality and enhance the overall developer experience.

+++ src/modules/reviews/services/place-reviews.service.ts
@@ -113,7 +113,7 @@ export class PlaceReviewsService {
       this.reviewRepository.findOne({ where: { id: reviewId, user: { id: userId }, place: { id: placeId } } }),
       this.placeRepository.findOne({ where: { id: placeId }, select: { id: true, rating: true, reviewCount: true } }),
     ]);
-    if (!review) throw new NotFoundException('Review not found');
+    if (!review || !place) throw new NotFoundException('Review not found');
 
     this.reviewRepository.merge(review, reviewDto);
     await this.reviewRepository.save(review);
@@ -133,7 +133,7 @@ export class PlaceReviewsService {
       this.reviewRepository.findOne({ where: { id: reviewId, user: { id: userId }, place: { id: placeId } } }),
       this.placeRepository.findOne({ where: { id: placeId }, select: { id: true, rating: true, reviewCount: true } }),
     ]);
-    if (!review) throw new NotFoundException('Review not found');
+    if (!review || !place) throw new NotFoundException('Review not found');
 
     await this.reviewRepository.remove(review);
